+++ src/main/kotlin/com/insyncwithfoo/pyrightls/configuration/application/ConfigurationPanel.kt
@@ -69,6 +69,10 @@ private fun Row.makeTaggedHintsInput(block: Cell<JBCheckBox>.() -> Unit) =
     checkBox(message("configurations.taggedHints.label")).apply(block)
 
 
+private fun Row.makeAutoImportCompletionsInput(block: Cell<JBCheckBox>.() -> Unit) =
+    checkBox(message("configurations.autoImportCompletions.label")).apply(block)
+
+
 @Suppress("DialogTitleCapitalization")
 internal fun configurationPanel(state: Configurations) = panel {
     // FIXME: The onInput() callbacks are too deeply nested.
@@ -104,16 +108,19 @@ internal fun configurationPanel(state: Configurations) = panel {
     }
     
     group(message("configurations.group.languageServer")) {
-        group(message("configurations.group.languageServer.client")) {
+        buttonsGroup(message("configurations.group.languageServer.client")) {
             row {
                 makeHoverSupportInput { bindSelected(state::hoverSupport) }
                 makeCompletionSupportInput { bindSelected(state::completionSupport) }
                 makeGoToDefinitionSupportInput { bindSelected(state::goToDefinitionSupport) }
             }
         }
-        group(message("configurations.group.languageServer.server")) {
+        buttonsGroup(message("configurations.group.languageServer.server")) {
             row {
                 makeTaggedHintsInput { bindSelected(state::taggedHints) }
+            }
+            row {
+                makeAutoImportCompletionsInput { bindSelected(state::autoImportCompletions) }
             }
             row(message("configurations.logLevel.label")) {
                 makeLogLevelInput { bindItem(state::logLevel.toNullableProperty()) }